requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
pydantic==2.4.2
pytest==7.4.2
pytest-cov==4.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0